Support for working around a failing server.
	Apparently, the common thing to do is check the status of a server when
performing a command against it, and trying the next server in the list when
the server is unavailable.

	My previous behavior was to queue a command against a server regardless of
its status such that when a server was available the command would be processed
as close to immediately as contention would allow, but when a server was
unavailable, the command would wait around for the server to come back up.

	The new behavior is something near the middle of this.  When a server is
unavailable, I'll try the next server in the list.  I repeat until I've
exhausted every known server and then just queue it for the one to which it
belonged in the first place.