Moved operation timeout into the connection factory.
It shouldn't change at runtime in a client, as the current model was
not thread safe.  In order for this to be dynamically reconfigurable,
the field either needs to be volatile or access to the value must be
synchronized.

By having the value be declared final and removing the mutators and
accessors, we can guarantee correctness.