Turning the timeouts way up during tests.
I manage to do something that slows stuff down in my full test plan
which leads to spurious failures.

This may be hinting at a bug of some type, or perhaps some bad
defaults, but this code isn't attempting to test timeouts so they
should never get in the way.