Read before writes and expect IllegalStateException.
These are two measures that are helping with the queue overflow
problems.

Firstly, the IllegalStateException is thrown whenever you attempted to
add to a queue that's full.  If that happens internally, I don't want
the IO thread to crash, so I add it to the normal ``expected
exceptions'' list.

Secondly, reading before writing helps keep the read buffer ready for
new data.  When writing first, the write will have to transition to a
read and may cause the read buffer to overflow.  It still may happen,
but by servicing the reads first I can at least get the complete ones
popped out before piling new ones in (since writes are almost always
smaller and likely to transition).