Killed off a slightly elusive bug that was causing sporatic cancellations.
Whenever a read returned in the binary protocol without having read
enough bytes to fill a header packet, an NPE would be fired that would
cause us to disconnect from the server and cancel all in-flight
operations.

This happened occasionally in one of my tests and was rather a pain to
track down.  It likely never affected anyone since I doubt anyone is
actually using the binary protocol anywhere today.