If we attempt to read when we don't have anything to read, reconnect.
This seems to happen in some strange server-error cases where the
server short-circuits the connection.  This situation has shown itself
when we've sent something too large into the server.  Precautions have
been taken to prevent that exact thing from happening again, but this
is a better way to handle an unexpected state.

I'm not sure how we end up still having data to receive when we've
already thrown away the ops, but tearing down the connection is a
clean way to recover from a bad state.