Some javadoc cleanup in MemcachedClient examples.
I don't think it'd be a change between Java 1.5 and Java 6, but in any
event I found the advanced example didn't compile without the changes
incorporated in the diff in the comments here.  Since binary protocol
is a feature, I thought this was a good place to show it off too.