Handle auth failures more gracefully; maximum failures.
This change will ensure that autentication failures do not
hang the thread adding an operation.  The default behavior
is to try to authenticate forever and whine to the log if
authentication continues to fail.

Optionally, a developer may define
net.spy.memcached.auth.AuthThreshold with an integer value
to set a maximum number of authentication attempts before
shutting down.

Authentication threads are now spawned and managed by an
AuthTreadMonitor class, ensuring that for any given
connection there can only be one outstanding AuthThread.

Change-Id: I0af4ab41d27239e7b555dca7d993e6228b7df5b8