Enhance MemcachedNode to know whether auth should happen.
When SASL authentication is requested of a MemcachedClient it should
always ensure that auth takes place before anything else on the queue
is processed.

After some experimentation, it seems The best way to handle this is to
extend the MemcachedNode interface to add some methods. This in turn
allows for the changes to be implemented in the TCPMemcachedNodeImpl
class.

There is special handling with the inputQueue and addOperation if
a reconnect occurs during authentication.

It's a bit harder at the BinaryMemcachedNodeImpl, though arguably
makes more sense there.  Since these are all pretty well hidden as an
implementation detail, it's okay for it to be here.

Who knows, maybe someday ascii will have SASL.  Probably not though.