Do not write timedout operations to the MemcachedNode.
This commit and related ones add support to an operation to
have new methods and a state of TIMEDOUT.  The intent is to
keep track of when an operation is created and if it either
times out due to a latch timeout expiration or it is found
to be already too old when thinking about sending the op
to the network, just consider it timed out then and there.

Note, object creation time is actually possibly quite a bit
after when the request is made, depending on how that request
is made.  Any number of things could have happened in
between with GC, JIT, scheduling, etc.

Also note that in order to avoid needing to rely on the latch
which is in a different thread, this allows us to keep track
of the creation time of the operation and check for whether or
not it has timed out via the isTimedOut() method on the
operation.

Doing this correctly and with as little API change as possible
required getting the default operation timeout down to the
MemcachedNode level.  That information was not previously known
to the Operation or the node, but simply relied on a latch.

Change-Id: I60228433bfa121ed031dd81fc05a9d65cae5bf20
Reviewed-on: http://review.membase.org/4204
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>