Increased default timeout to 2500ms.
The increase of the timeout to this seemingly high value is due to
a few findings.

First, by default, garbage collection times may easily go over 1sec.
Testing with simple toy tests shows this quite clearly, even on
systems with lots of CPUs and a decent amount of memory.  Of course,
much of this can be controlled with GC tuning on the JVM.  With the
hotspot JVM, look to this whitepaper:
http://java.sun.com/j2se/reference/whitepapers/memorymanagement_whitepaper.pdf

Testing showed the following to be particularly useful:
-XX:+UseConcMarkSweepGC -XX:MaxGCPauseMillis=850

There is a CPU time tradeoff for this.

Even with these, testing showed some 1 second timeouts when GCs near a
half a second.  To use this software though, we shouldn't expect people
to have to tune the GC, so raising the default seems like the
right thing to do.

Second, many systems use spymemcached on virtualized or cloud environments.
The processes running there do not have any guarantee of execution
time.  It'd be really unlikely for a thread to be starved for more than
a second, but it is possible and shouldn't make things stop.  Raising this
a bit will help.

Third, and perhaps most importantly, most people run applications on
networks that do not offer any guarantee around response time.  If
the network is oversubscribed or even minor blips occur on the network
can cause TCP retransmissions.  While many TCP implementations ignore
it, RFC 2988 specifies rounding up to 1sec when calculating
TCP retransmit timeouts.  Blips will occur, and rather than force
this seemingly synchronous get to timeout, it may be better to
just wait a bit longer by default.

Change-Id: Ie53ca774458466d9a2e6f70e65ea6663699a9f6f
Reviewed-on: http://review.membase.org/4207
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Michael Wiederhold <mike@membase.com>
Tested-by: Dustin Sallings <dustin@spy.net>