Increase the maximum size allowed.  Issue 106.
Though memcached still has a 1MiB size by default (overridable
with a flag), other engines may have their own maximum.  Membase
for instance has a 20MiB maximum.

Note that the size enforced by the transcoder can be overridden
through it's ctor.  The reason for making this change is to
create a new, larger default.

At least for the remainder of the 2.6 series, rather than do
deep surgery on this issue, just bump up the maximum.

Change-Id: I3e7bb000b12a63b4a299cb8b79a3151fbfc91615
Reviewed-on: http://review.membase.org/6074
Reviewed-by: Michael Wiederhold <mike@membase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Matt Ingenthron <matt@northscale.com>