Added TEMP_FAIL handling to set operations
TMP_FAIL is sent to the client when the server is out of memory.
The current way of handling this error is to reconnect, but this
is no the correct way to handle a this error since the server will
be available soon.

There are other commands that should handle this as well, but right
now I am only submitting a fix for the set operation since this is
the operration that causes this error most frequently. We plan on
adding support for all known Membase commands in a future release.

Change-Id: I8099dd7c85a45c81667e3a566995995dd3780b32
Reviewed-on: http://review.couchbase.org/6751
Tested-by: Michael Wiederhold <mike@membase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Dustin Sallings <dustin@spy.net>