ASCII get operations now return a false operation status on failure
get in ASCII always returns END at the end of the result from
memcached. In the current implementation of spy seeing END caused
spy to return operation success. We now check to see if a value
is returned before END and if one if not we return a failure
for OperationStatus of NOT_FOUND.

Change-Id: Ia84df9398e258ae1edba7a11dee55a4c265414d9
Reviewed-on: http://review.couchbase.org/7040
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>