Add visibility into operations (status).
This commit allows the functionality for users to issue requests
and then check to see whether or not they succeeded. They can also
now access the error message that the server sends back to them
so they can see why the server rejected their request. Operation
status's also take into account operation timeouts and exceptions
so if an operation times out the operation says so. Also if an
exception is thrown while processing the operation the operation
status reports the message from that exception.

Change-Id: I62af7450cf6cd1c9d1bf171e5063b3d8a1c919ce
Reviewed-on: http://review.couchbase.org/7044
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>