Add visibility into operations (key)
This commit adds the ability for the user to be able access
the key that the operation asked for if one is present. This allows
users to be able to easily resend an operaiton on a given key since
they will no longer have to keep key-operation data structures
around in the case of failures.

Change-Id: I2db360abbc0e4577526ac057a4945eb55d0dbd10
Reviewed-on: http://review.couchbase.org/7045
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>