Made SyncGetTest failures less sporadic
Moved the code that sets the items for this test so that it could
use a MemcachedClient that had a regular timeout. This allows us to
avoid a possible failure resulting from using the connection that
has a short timeout. Now only the get (which is surposed to fail)
uses this short timeout connection

Change-Id: Ic085218bcd6fa53348f70eae13e219a9882b6442
Reviewed-on: http://review.couchbase.org/7646
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>