Changed the value size of items used in LongClientTest
The 32k value used previously worked fine when we were only testing
clients against localhost, but in the future we want to be able to
test spymemcached against servers on the network. This high value
along with the large number of multi-gets done in this test causes
failures due to not being able to read and write data fast enough.
As a result timeouts were caused.

The value has been lowered to something more reasonable for now, but
in the future we should devise a test that will be able to test spy's
ability to handle multi-gets with large values over a network.

Change-Id: I94c7d61bfb6dd76abf4afccd0796dca99fd4bd0a
Reviewed-on: http://review.couchbase.org/7647
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>