Added a command line parameter for specifying server type
The user naow has the ability to specify the server type that
they are testing against by adding the parameter 'SPYMC_SERVER_TYPE'
to the command line. The default type is memcached and there are
three valid types, memcached, membase, and couchbase. An example
of how to use the parameter is below

buildr test SPYMC_SERVER_TYPE="membase"

Change-Id: I8d4750053cf52734c703bfe80d8c2d42ca353bb0
Reviewed-on: http://review.couchbase.org/7651
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>