Refactored Operations to improve correctness of vbucket aware ops
Previously we had operations that implemented classes that they
should not have been implemented. For example, there were ASCII
operations that implementing the VBucketAware interface. Since
ASCII operations must go through moxi and don't contain vbucket
numbers there is no reason for them to implement this interface.

This commit also contains the addition of SingleKeyOperationImpl
and MultiKeyOperationImpl which subclass OperationImpl and are
superclasses to binary operations that contain one or more keys
correspondingly. This abstraction allows these type of operations
to deal with vBucket awareness and the amount of keys the operation
contains in different ways.

Change-Id: Ic7d002ec8e77e2fccc31987fcb1cba1ca14715da
Reviewed-on: http://review.couchbase.org/7801
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>