Ensure nodesMap updates are safe when topology changes.
This change much more closely (perhaps too defensively)
couples the Config and the list of nodes.  It also ensures
that the Map<String, MemcachedNode> is less prone to
concurrency problems and fixes a concurrency problem which
allowed nodes to be removed when receiving a not-my-vbucket
response.

This indicates we need to make some changes to how the entire
object model and config list are handled, but just to quickly
fix a race found, tie together the config and node list when
locating the right node.

Testing for this is rather manual at the moment.

Change-Id: Ida002fd1d510d20c432e77e01eefbc530d3b34b1
Reviewed-on: http://review.couchbase.org/8195
Tested-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>