Tap streams now pause every 10,000 messages.
In the previous code I was sending acks back to the server
immediately and this kind of behavior could cause a heap overflow
for users who start tap streams but don't process the messages. Now
the ack messages are queued with the responses and only sent back
to the server after all mutations sent before the ack was recieved
are processed.

Change-Id: I0c7c84c2ae5c9cc68d4e9c6f925f9ac6885fae6b
Reviewed-on: http://review.couchbase.org/8114
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>