Encode with commons codec more correctly.
Apache Commons Codec had introduced a change in 1.4 that had it
inadvertantly adding extra CRLF to encode responses.  We had worked
around this issue when writing the initial HTTP authorization
code, but it turns out commons codec has reverted to the behavior
in 1.3 with the release of 1.5.

For more info, see https://issues.apache.org/jira/browse/CODEC-89

This change updates commons codec, adds a test to verify correct
behavior, and addresses potential error handling issues with the
character set.

Change-Id: I655adc5090249180c3e75fbd35036b15252d093f
Reviewed-on: http://review.couchbase.org/8125
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>