Added the ability for Spy to handle errors in views
If a node is down or an item is unavailable then Couchbase
will return a view response that contains a mix of errors
and successes. Previously Spy could only handle responses
that contained only successes. This commit adds the ability
to handle a mix of errors and successes returned in the same
view result.

Change-Id: I2795962b8bf56a81f7a2c9ebacf4c90aaad4ce3c
Reviewed-on: http://review.couchbase.org/8461
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>