SPY-8: Added a limit on how many gets can be optimized
We have a limit on the amount of sets that can be optimized so we
should have a limit on how many gets can be optimized as well. This
way people don't run into any issues with optimizing too many get
requests and having the client timeout as a result of the server
being too slow trying to fill all of the requests in the multi-
get.

The max number of gets that can be optimized into a single
request is 4096.

Change-Id: I23c107ca3459e3386270c53f9ce20450d7ee8830
Reviewed-on: http://review.couchbase.org/8599
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>