Added interface for view rows called ViewRow
Each row type that can be sent back from a view now has
getter functions for every possible field. If a field doesn't
exist for that view row an UnsupportedOperationException is
thrown. This new interface is now the only user facing object
that will be seen when dealing with view rows and as a result
makes the code simpler to use

Change-Id: Ie8a9548846932d2a11570a1ed2b7a5fdac1d3723
Reviewed-on: http://review.couchbase.org/8998
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>