Log warnings when retrying due to not my vbucket.
The existing behavior with not my vbucket errors prior to this
change would warn when it receives an error to the log, but it
would not followup that warning with any indication of what it
did with the operation.  This can cause confusion end users
are logging what's happening, so give the logging some symmetry
by logging any retries due to not-my-vbucket.

This was brought over from the 2.7 refresh branch manually,
owing to the formatting change making it hard to cherry-pick.

Change-Id: I233c3c854adcfe4886e879e69f568807def595a7
Reviewed-on: http://review.couchbase.org/8997
Tested-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>