Operations can't timeout when writing to the write buffer.
We implemented a mechanism for operations to be able to timeout
while waiting to be written to a socket, but failed to take into
account the fact that an item can timeout while in the middle
of writing to that buffer.

This commit adds a new OperationState called WRITE_QUEUED that
specifies that the operation is waiting to be written to the
write buffer. When an operation begins writing to the write
buffer it goes into the WRITING state. When an operation is in
the WRITING state it cannot time out

Change-Id: Ib4ed9e1764c31e9270f79206b11574d98638e989
Reviewed-on: http://review.couchbase.org/9424
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>