SPY-125: Significant performance issue large number of sets
We now add an operation to the read queue when we transistion it
into a WRITING state. This guarentees that the operation is added
to the read queue once. As a result we no longer have to check to
see if the operation has already been added to the read queue

Change-Id: I0b21d3e1ac7ff7792afabc57c1a8a26c6b757c12
Reviewed-on: http://review.couchbase.org/9425
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>