Improved performance of write queue processing during timeouts
There were two major issues here. First if there are a lot of
timeouts in a row then we processing the write queue we removed
them only one at a time. We now are able to process the entire
write queue each time the handleIO function is called. The
second issue was that it is possible to send timed out requests
to the server. This would happen if there was one not timed out
request followed by some timed out requests. Since in the while
loop we don't check to see if an operation is timed out (or
cancelled) we would send it anyways. This commit addresses both
of these issues by adding a getNextWritable() op which only
operations that should be written. If the write queue is empty
it returns null.

Change-Id: I427e5a47cf2e8ad89e3f59c895f3ef8e0ad36550
Reviewed-on: http://review.couchbase.org/9429
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>