Change getBytes() to getData() in CASOperation
The getBytes function was incorrectly named in the CASOperation
class for two reasons. First the function actually gets the
data field from a message so getData() is a better name. Second,
in the StoreOperation class and other classes the name of the
function is getData() so this improves consistency. This issue
was notice when I was looking at the StoreOperation and saw that
this class had two different functions that did exactly the same
thing.

Change-Id: I4d0c95f309a6f1bb1ccde52ffdf6e9ab1958bebe
Reviewed-on: http://review.couchbase.org/9474
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>