SPY-50: Set viewmode programatically
We can now set properties programatically and I allowed the
properties file to be found in the classpath (which I thought
I did before, but I guess I didn't). There is now a heirarchy
for how this parameter can be defined. The list is below from
highest priority to lowest:

1. Set in user code - System.setProperty(2)
2. Set on the command line
3. Set in the cbclient.properties file

Change-Id: I1dc91641016fa31e5cdd51f5bba85bd7cd68bd45
Reviewed-on: http://review.couchbase.org/9462
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>