Made cmd variable a byte for binary operations
This variable should have never been an int since it is defined in
the protocol as a byte. The reason for the change is that when we
have opcode that is greater have 127 the value of a byte and an int
are different (Ex. int = 128 byte = -127). This was causing an
assertion to fail, in this case for the getl command (0x94).

Change-Id: Ia6b32a4af94ed2e8dc973237132f3fd291655634
Reviewed-on: http://review.couchbase.org/9476
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>