SPY-37 & SPY-38: Fixed redistribution performance issue
This issue was caused by a O(n^3) algorithm used when redistributing
operations among nodes. It turns out that one of the loops was not
necessary because the clone operation is guaranteed to return a
list of single key operations. Multi-key operations are never
returned as part of this list.

The fix involves removing the one of these loops. which frequently
queued duplicate operations for transmission. Since the
operations are all single key, the list returned by the call
getKeys() is guaranteed to have only one operation.

Change-Id: Id8cf17132bff06dba78935cbb8b81d83f58b4bc2
Reviewed-on: http://review.couchbase.org/9514
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>