Send an ack for all tap opaque messages
Previously we ignored tap opaque and no-op messages. This caused
us to hang tap streams in some cases. One example is when doing a
tap dump. The server would send an opaque message and wait for an
ack after transmitting all of the data. Since we ignored the opaque
message the server would wait for an ack that would never be sent.
This would cause the connection to hang.

Change-Id: If8fc5e7f8b6aeb41ed68c2acb03fba5b7e503836
Reviewed-on: http://review.couchbase.org/9526
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>