Fixed testOverflowing ReadQueue in QueueOverflowTest
The issue with this test failing inconsistently has to do with
how many operations are in the input queue when the first
queue overflow occurs. The queue size is 5 and if there are more
than 5 operations in the input queue then this will cause a second
queue overflow when we try to do our final operation. This fix
corrects the error by making sure that we drain all of the
operations from the write queue and therefore make sure that there
is actually space for the last operation.

Change-Id: I1e793514e8b1df1dbcdb133fba7cec837994b6d9
Reviewed-on: http://review.couchbase.org/9613
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>