Add compatibility with netty 3.2.0+.
Netty 3.2.0+ (NETTY-281) changed HttpMessage.setHeader(String, String) to
setHeader(String, Object), which causes NoSuchMethodErrors when spymemcached is
used with such a newer netty version. To prevent spymemcached users from other
library incompatibilities the netty dependency is not upgraded, but the
appropriate setHeader method is selected and invoked at runtime. This was
discussed on the mailing list:
http://groups.google.com/group/spymemcached/browse_thread/thread/4b9f0638bbcc96fb

Change-Id: I1d2e7c4a3e6db6648d175364950eab6076a45e5f
Reviewed-on: http://review.couchbase.org/9852
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>