Fixed handling of TAP flags.
The request and response messages for flags were not following the defined
protocol for their data size.  Also, the logic in the setting of the flags
was incorrect, meaning it wasn't actually possible to determine which
flags were set.

A test that verifies correct behavior will be added in a later commit
that introduces test coverage for the ResponseMessage class.

Change-Id: I20ae1b176cb3bc764e993ab98254b15f69cba9ea
Reviewed-on: http://review.couchbase.org/13995
Reviewed-by: Raghavan N. Srinivas <raghavan.srinivas@gmail.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>