Create an ErrorCode class for error codes
This allows us to use error codes in other places than just
operations. This class defines all of the Memcached error codes
as well as two client specific codes (timed out and canceled). The
reason for client specific error codes is that is gives the user
something definitive to test against when error cases happen.

Change-Id: I6b4b21ec8448444103d8e73b440df6085bdc1e33
Reviewed-on: http://review.couchbase.org/15948
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>