Revert ErrorCode and OperationStatus changes.
This is a squashed revert of two separate commits. After
disabling optimization, I still found failing tests. I had
thought it was related at first, but after doing a bisect
it clearly identified the "Create an ErrorCode class" as
the culprit. In combination with the OperationState
correctly being volatile, it would regularly fail in the
midst of bulk operations.

This commit reverts:
"OperationStatus now has an error code"
 commit cdc26cdfaeecc63714c1edcf297d6ffc2c9b8554 and
"Create an ErrorCode class for error codes"
 commit 82ad6cb4629a2c98fdebc55ad10f058cd2767ff9.

Change-Id: Id6e04ed1787e140c5072a78cece682c45fe06313
Reviewed-on: http://review.couchbase.org/18683
Reviewed-by: Raghavan N. Srinivas <raghavan.srinivas@gmail.com>
Tested-by: Raghavan N. Srinivas <raghavan.srinivas@gmail.com>