[WIP] Add touch to the ASCII protocol.
Touch is defined for the ASCII protocol in memcached, but not implemented in
spymemcached.  This commit adds the missing touch functionality.

The ant test suite as described in README.markdown does not pass out of the
box. Details of my local setup:

Mac OS X 10.8.3
Memcached 1.4.15
Spymemcached 2.8.12

Tests that fail as reported by ant:

[junit] Test net.spy.memcached.AsciiClientTest FAILED (timeout)
[junit] Test net.spy.memcached.AsciiIPV6ClientTest FAILED (timeout)
[junit] Test net.spy.memcached.BinaryClientTest FAILED (timeout)
[junit] Test net.spy.memcached.BinaryIPV6ClientTest FAILED (timeout)
[junit] Test net.spy.memcached.QueueOverflowTest FAILED
[junit] Test net.spy.memcached.TapTest FAILED