SPY-133: Add additional metrics support.
This changeset adds basic metrics for performance measurement and
debug support. It is optional and can be enabled through the factory
or via a system property.

The collected values can be exported through:
 - The Console
 - JMX
 - CSV
 - SLF4J

These output reporters can also be configured through properties, by
default the console reporter will be used.

Change-Id: If206f0be0c0bd6c6ff11d63f1be25debab2e524e
Reviewed-on: http://review.couchbase.org/28575
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>