SPY-141: Do not assume positive CAS.
This changeset makes sure that when loading the CAS value, there
is no assumption of the actual value done on the client side.
Especially, -1 is a valid CAS and can cause issues when running
with -ea in production (since the callback is executed on the
IO thread, it can die for no reason).

Change-Id: Ic944bca9c7e005fb793edc0ddb3a51674f2d2be0
Reviewed-on: http://review.couchbase.org/29584
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>