CCCP: Implement NOT_MY_VBUCKET handling customization.
This changeset allows MemcachedConnection subclasses to make use
of any RETRY response that arrives, handing over its body. For
example, in Couchbase this can be utilize to get the new server
configuration and pass it over to the configuration management
subsystem.

Note that all this is happening on the IO thread, so whatever
method subclasses it should make sure to not block the thread
at all.

Change-Id: Ia71874207b5a213074e2ca525aa3eb37ff80e24b
Reviewed-on: http://review.couchbase.org/30448
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>