SPY-146: Use correct length of key in observe.
This changeset makes sure that the correct length of the key is used in the
protocol. Since UTF-8 chars like the pound sign (currency) are represented
in two bytes but are one character in a string, the length of the key needs
to be determined after it is converted into a byte array.

Change-Id: I48b68127f384e3bcb963a5df1cddd5b5c2d5ca7a
Reviewed-on: http://review.couchbase.org/30701
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>