SPY-149: Notify Listeners after Complete.
This changeset honors the fact that setting the status and actually
completing the latch can happen at two different points in time.
Without this change, effects can be observed that a listener is called
but the latch has not been counted down yet, leading to weird effects.

Also see https://github.com/mathieuancelin/play2-couchbase/issues/33
for a dependent issue.

Change-Id: I99686c212d8610299f78078b28d31b7891da8d1c
Reviewed-on: http://review.couchbase.org/31714
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>