SPY-153: Adding a typesafe StatusCode for OperationStatus instances.
This makes it possible to add easier checks on futures, while not
having to rely on string matching in the status message. This is
an addition, so old behavior will not be changed, its just more
convenient and can be better documented (it does not directly rely
on a server response).

Change-Id: Ie416e317d5fd21af2c3efb177b08b68dd8c45639
Reviewed-on: http://review.couchbase.org/33405
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>