SPY-154: Also RETRY on binary NMVB response with bulk gets.
This changeset makes sure that even when one of the individual get
operations in the bulk request fails with a "Not My VBucket", the
single get operation gets retried.

To make this work, the underlying bulk operation is changed slightly
to be notified that it has to wait for N more operations before
notifying the listeners and counting down the latch.

Change-Id: I6413233a3d0289ab728312f8f4edab5df96a0081
Reviewed-on: http://review.couchbase.org/33700
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>