SPY-127: Optimize StringUtils performance.
This changeset optimizes the individual StringUtils methods, especially
the isJsonObject method. Before, if no proper JSON was detected, a final
check against an Integer was performed. If no integer was found, a very
expensive exception was thrown, killing performance.

This check has been replaced with a faster regex. Also, other parts
of the methods have been refactored to be more performant.

Change-Id: I7616d2b042896aa76d73d536a3dde1c998990d9b
Reviewed-on: http://review.couchbase.org/34184
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>