SPY-158: Use Milliseconds instead of Seconds for "max reconnect" delay.
Motivation
----------
Currently, when a reconnection needs to happen against a node, a incremental
delay for subsequent reconnect tries is selected. Because of the wrong time
format used (seconds instead of milliseconds), the delay is always capped
at 30ms, leading to immediate reconnects with no backoff effect.

Modifications
-------------
For backwards compatibility, the "second" time unit is kept in the factory,
and instead in the MemcachedConnection the time is converted to milliseconds.

Result
------
A proper reconnect delay going from 2^0 seconds up to 2^4 seconds, peaking at
30 seconds by default (as always changeable through the factory)

This changeset makes sure that

Change-Id: I4a348f50e41451a6542b4b0263b915990730c3f8
Reviewed-on: http://review.couchbase.org/34728
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>