SPY-159: Make the auth latch timeout configurable.
Motivation
----------
Currently, the auth latch timeout is fixed to one second and not changeable.
In certain scenarios (for example with CRAM-MD5 where more roundtrips need
to be issued), this is too short. Making it configurable allows to adjust the
setting to latencies in target environments.

Modification
------------
Expose a "authWaitTime" setting in the factory and the builder to allow set
the value in milliseconds. It still defaults to one second for backwards
compatibility.

Result
------
The auth timeout can be modified as needed in production deployments.

Change-Id: Id06f4ede00402688c00a283bcff82cde25a7a205
Reviewed-on: http://review.couchbase.org/34846
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>